requirements.txt
requests
beautifulsoup4
paho-mqtt
python-dotenv